backend/requirements.txt
aiohttp
aiohttp_cors
motor
python-dotenv
mypy
